+++ src/userspace/libc/stdlib/htoi.c
@@ -5,9 +5,9 @@
 
 int htoi(char *str)
 {
-	u8 s_str = strlen(str);
+	u32 s_str = strlen(str);
 
-	u8 i = 0;
+	u32 i = 0;
 	int ret = 0;
 	for (; i < s_str; i++) {
 		char c = str[i];
