+++ src/userspace/libc/string/strcati.c
@@ -3,8 +3,8 @@
 
 void strcati(char *dest, char *orig)
 {
-	u8 s_orig = strlen(orig);
+	u32 s_orig = strlen(orig);
 	strdisp(dest, (int)s_orig);
-	for (u8 i = 0; i < s_orig; i++)
+	for (u32 i = 0; i < s_orig; i++)
 		dest[i] = orig[i];
 }
