+++ src/userspace/libc/string/strdup.c
@@ -5,7 +5,7 @@
 
 char *strdup(char *orig)
 {
-	u8 s_orig = strlen(orig);
+	u32 s_orig = strlen(orig);
 	char *ret = (char *)malloc(s_orig + 1);
 	strcpy(ret, orig);
 	return ret;
