+++ src/userspace/libc/string/strlen.c
@@ -1,8 +1,8 @@
 #include <stdint.h>
 
-u8 strlen(char *str)
+u32 strlen(char *str)
 {
-	u8 len = 0;
+	u32 len = 0;
 	while (str[len])
 		len++;
 	return len;
